Projeto terminado

+++ Login/Quiz.py
@@ -214,12 +214,10 @@ class BrainBuster:
 class QuizBuster:
     
     def __init__(self): 
-        
         self.qa_list = []
         self.janela_principal()
 
     def janela_principal(self):
-        
         self.janela = tk.Tk()
         self.janela.title('Menu')
         self.janela.state('zoomed')
@@ -230,7 +228,7 @@ class QuizBuster:
         b_r = tk.Button(self.janela, text='Quiz', height=2, width=15, border=3, borderwidth=3, command=self.quiz)
         b_r.place(rely=0.45, relx=0.35)
 
-        b_l = tk.Button(self.janela, text='Perguntas', height=2, width=15, border=3, borderwidth=3, command=self.lista)  # Chame o método corretamente
+        b_l = tk.Button(self.janela, text='Perguntas', height=2, width=15, border=3, borderwidth=3, command=self.lista)
         b_l.place(rely=0.45, relx=0.45)
 
         b_i = tk.Button(self.janela, text='Informações', height=2, width=15, border=3, borderwidth=3)
@@ -245,23 +243,29 @@ class QuizBuster:
         self.janela.mainloop()
 
     def quiz(self):
-        
         self.janela.destroy()
         self.janela1 = tk.Tk()
         self.janela1.title('Quiz')
         self.janela1.state('zoomed')
 
-        self.panel = tk.Label(self.janela1, text='BrainBuster', font=('Georgia', 23))
-        self.panel.place(rely=0.3, relx=0.44)
+        self.selecionar_pergunta()  # Obtém uma nova pergunta
 
-        b_r = tk.Entry(self.janela1, width=15, font=('Georgia', 13))
-        b_r.place(rely=0.45, relx=0.445)
+        if self.f: 
+            self.p_t = tk.Label(self.janela1, text=self.f, font=('georgia', 20))
+            self.p_t.place(relx=0.5, rely=0.2, anchor='center')
+        else:
+            messagebox.showinfo('Sem perguntas', 'Não tem pergunta registrada!')
+            self.janela1.destroy()
+            return
 
-        self.mo = tk.Button(self.janela1, text='Enviar', border=3, borderwidth=3, width=6)
-        self.mo.place(rely=0.45, relx=0.55)
+        self.r_e = tk.Entry(self.janela1, font=2, border=3, borderwidth=3, width=15)
+        self.r_e.place(relx=0.5, rely=0.45, anchor='center')
 
-        self.pe = tk.Button(self.janela1, text='Pular pergunta', border=3, borderwidth=3)
-        self.pe.place(rely=0.5, relx=0.465)
+        b_e = tk.Button(self.janela1, text='Enviar', border=3, borderwidth=3, width=5, command=self.revisar_resposta)
+        b_e.place(relx=0.58, rely=0.45, anchor='center')
+
+        b_p = tk.Button(self.janela1, text='Pular pergunta', border=3, borderwidth=3, width=10, command=self.atualizar_pergunta)
+        b_p.place(relx=0.5, rely=0.5, anchor='center')
 
         b_r = tk.Button(self.janela1, text='Retornar', height=2, width=17, border=3, borderwidth=3, command=lambda: self.retornar(self.janela1))
         b_r.place(rely=0.94, relx=0.02)
@@ -269,8 +273,46 @@ class QuizBuster:
         b_s = tk.Button(self.janela1, text='Sair', height=2, width=17, border=3, borderwidth=3, command=lambda: self.sair(self.janela1))
         b_s.place(rely=0.94, relx=0.9)
 
-    def lista(self):
+        self.janela1.mainloop()
+
+    def atualizar_pergunta(self):
         
+        messagebox.showinfo('Pulando pergunta', "Atualizando pergunta...")  # Mensagem de depuração
+        self.selecionar_pergunta()  # Seleciona uma nova pergunta
+
+        if self.f:
+            
+            self.p_t.config(text=self.f)  # Atualiza o texto do label
+        else:
+            messagebox.showinfo('Sem perguntas', 'Não há mais perguntas registradas!')
+
+    def revisar_resposta(self):
+        resposta_usuario = self.r_e.get()  # Obtém o texto da entrada
+        
+        if resposta_usuario.strip().lower() == self.g.lower():  # Comparação sem diferenciar maiúsculas
+            messagebox.showinfo('Certa resposta', 'Você acertou a resposta!')
+        else:
+            messagebox.showinfo('Resposta errada', 'Você errou a resposta.')
+
+    def selecionar_pergunta(self):
+        
+        with conexao.cursor() as cursor:
+        
+            sql = 'SELECT Perguntas, Respostas FROM estudos ORDER BY RAND() LIMIT 1'  # Aleatório para novas perguntas
+            cursor.execute(sql)
+            resultado = cursor.fetchone()
+
+        if resultado:
+            
+            self.f = resultado[0].strip()  # Acessa a pergunta
+            self.g = resultado[1].strip()  # Acessa a resposta
+        
+        else:
+            
+            self.f = None  
+            self.g = None
+
+    def lista(self):
         self.janela.destroy()
         self.janela2 = tk.Tk()
         self.janela2.title('Registro de Perguntas e Respostas')
@@ -278,22 +320,18 @@ class QuizBuster:
 
         tk.Label(self.janela2, text='Registro de Perguntas e Respostas', font=('Arial', 16)).pack(pady=10)
 
-        # Campo para pergunta
         pe = tk.Label(self.janela2, text='Digite sua pergunta:')
         pe.pack(pady=5)
-        
-        self.question_entry = tk.Entry(self.janela2, width=40)
-        self.question_entry.pack(pady=5)
 
-        # Campo para resposta
+        self.p_e = tk.Entry(self.janela2, width=40)
+        self.p_e.pack(pady=5)
+
         tk.Label(self.janela2, text='Digite sua resposta:').pack(pady=5)
         self.answer_entry = tk.Entry(self.janela2, width=40)
         self.answer_entry.pack(pady=5)
 
-        # Botão para registrar
         tk.Button(self.janela2, text='Registrar Pergunta e Resposta', command=self.registra_pergunta).pack(pady=10)
 
-        # Listbox para exibir perguntas e respostas
         self.listbox = tk.Listbox(self.janela2, width=50, height=10)
         self.listbox.pack(pady=10)
 
@@ -305,71 +343,64 @@ class QuizBuster:
         b_s = tk.Button(self.janela2, text='Sair', height=2, width=17, border=3, borderwidth=3, command=lambda: self.sair(self.janela2))
         b_s.place(rely=0.94, relx=0.9)
 
-        self.carregar_perguntas()  # Carrega as perguntas ao abrir a janela
-
+        self.carregar_perguntas() 
+        
     def deletar(self):
         try:
-            selected_index = self.listbox.curselection()[0]  # Obtém o índice selecionado
-            selected_item = self.listbox.get(selected_index)
-            item_id = int(selected_item.split('|')[0].split(': ')[1])  # Extrai o ID do item selecionado
+            selected_index = self.listbox.curselection()  
+            if selected_index:  # Verifica se há algo selecionado
+                selected_index = selected_index[0]  
+                selected_item = self.listbox.get(selected_index)
+                item_id = int(selected_item.split('|')[0].split(': ')[1])  
 
-            # Remover do banco de dados
-            with self.conexao.cursor() as cursor:
-                sql = 'DELETE FROM estudos WHERE id = %s'
-                cursor.execute(sql, (item_id,))
-                self.conexao.commit()
+                with conexao.cursor() as cursor:
+                    sql = 'DELETE FROM estudos WHERE id = %s'
+                    cursor.execute(sql, (item_id,))
+                    conexao.commit()
 
-            # Remove do Listbox
-            self.listbox.delete(selected_index)
-            messagebox.showinfo('Sucesso', 'Pergunta e resposta apagadas com sucesso!')
+                self.listbox.delete(selected_index)
+                messagebox.showinfo('Sucesso', 'Pergunta e resposta apagadas com sucesso!')
+            else:
+                messagebox.showwarning('Aviso', 'Por favor, selecione uma pergunta e resposta para apagar.')
         
-        except IndexError:
-            messagebox.showwarning('Aviso', 'Por favor, selecione uma pergunta e resposta para apagar.')
         except pymysql.Error as error:
             messagebox.showerror('Erro', f'Ocorreu um erro ao apagar a pergunta: {error}')
         
     def registra_pergunta(self):
-        question = self.question_entry.get()
-        answer = self.answer_entry.get()
+        self.question = self.p_e.get()
+        self.answer = self.answer_entry.get()
     
         try:
             with conexao.cursor() as cursor:
-                if not question or not answer:
+                if not self.question or not self.answer:
                     messagebox.showwarning('Aviso', 'Por favor, preencha tanto a pergunta quanto a resposta.')
-                else:
-                    sql = 'INSERT INTO estudos(Perguntas, Respostas) VALUES(%s, %s)'
-                    cursor.execute(sql, (question, answer))
-                    conexao.commit()
+                    return
                 
-                    # Adicione a pergunta e resposta ao Listbox
-                    self.listbox.insert(tk.END, f'Pergunta: {question} | Resposta: {answer}')
-                
-                    self.question_entry.delete(0, tk.END)
-                    self.answer_entry.delete(0, tk.END)
-                    messagebox.showinfo('Sucesso', 'Pergunta e resposta registradas com sucesso!')
-                
+                sql = 'INSERT INTO estudos(Perguntas, Respostas) VALUES(%s, %s)'
+                cursor.execute(sql, (self.question, self.answer))
+                conexao.commit()
+                self.listbox.insert(tk.END, f'Pergunta: {self.question} | Resposta: {self.answer}')
+                self.p_e.delete(0, tk.END)
+                self.answer_entry.delete(0, tk.END)
+                messagebox.showinfo('Sucesso', 'Pergunta e resposta registradas com sucesso!')
+    
         except pymysql.Error as error:
             conexao.rollback()
             messagebox.showerror('Erro', f'Ocorreu um erro ao registrar a pergunta: {error}')
-
+        
     def carregar_perguntas(self):
         try:
             with conexao.cursor() as cursor:
-                sql = 'SELECT Perguntas, Respostas FROM estudos'
+                sql = 'SELECT id, Perguntas, Respostas FROM estudos'
                 cursor.execute(sql)
-                resultados = cursor.fetchall()
+                resultado = cursor.fetchall()
             
-                # Limpa o Listbox antes de carregar os dados
                 self.listbox.delete(0, tk.END)
-                for pergunta, resposta in resultados:  # Adiciona cada resultado ao Listbox
-                    self.listbox.insert(tk.END, f'Pergunta: {pergunta} | Resposta: {resposta}')
+                for id, pergunta, resposta in resultado:
+                    self.listbox.insert(tk.END, f'ID: {id} | Pergunta: {pergunta} | Resposta: {resposta}')
                 
         except pymysql.Error as error:
             messagebox.showerror('Erro', f'Ocorreu um erro ao carregar as perguntas: {error}')
-            
-    def mostrar(self):
-        self.panel.place(relx=0.3, rely=0.44)
-        
 
     def sair(self, janela_atual):
         r = messagebox.askquestion('Fechar a janela', 'Você realmente deseja fechar a janela?')
@@ -380,6 +411,6 @@ class QuizBuster:
         r = messagebox.askquestion('Retornar janela', 'Você realmente deseja retornar ao menu?')
         if r.lower() == 'yes':
             janela_atual.destroy()
-            self.janela_principal()  # Chama o método para reiniciar a janela principal
+            self.janela_principal()
     
 BrainBuster()
